+++ src/hyperlight_host/src/hypervisor/regs.rs
